+++ src/main/java/engine/PostMoveValidator.java
@@ -1,4 +1,11 @@
+package engine;
+
 import java.util.List;
+
+import engine.entities.Board;
+import engine.move.Move;
+
+package engine;
 public class PostMoveValidator {
     public List<Move> validMoves;
     /**
