+++ src/main/java/engine/entities/BoardCreator.java
@@ -1,3 +1,5 @@
+package engine.entities;
+
 /**
  * A creator for the "typical" Board instance.
  * The created Board object will have new pieces placed
