+++ src/main/java/engine/entities/GameCreator.java
@@ -1,3 +1,5 @@
+package engine.entities;
+
 /**
  * A creator for a new Game instance, representing
  * a chess match in which no moves have been made.
