+++ src/main/java/engine/entities/PieceFactory.java
@@ -1,3 +1,5 @@
+package engine.entities;
+
 public class PieceFactory {
     public PieceFactory() {}
