+++ src/main/java/engine/entities/Player.java
@@ -1,3 +1,5 @@
+package engine.entities;
+
 import java.util.List;
 import java.util.ArrayList;
