+++ src/main/java/engine/entities/copy/BoardCopier.java
@@ -1,3 +1,7 @@
+package engine.entities.copy;
+
+import engine.entities.Board;
+
 /**
  * A deep cloner for Board instances.
  */
