+++ src/main/java/engine/movegen/ValidMovesGen.java
@@ -1,4 +1,8 @@
+package engine.movegen;
+
 import java.util.List;
+import engine.move.Move;
+
 public class ValidMovesGen {
     public List<Move> validMoves;
 }
